Fix retry logic
The recursive call to `apply_rule` had the argument order wrong.
It passed `try_number+1` as the argument for `tries`.
The effect of that was to change `tries` to `1` on the first
recursive call. Thus only one try was ever attempted.

This commit fixes the argument order thus enabling retries.